+++ benchmark/Benchmark/Type.hs
@@ -16,14 +16,15 @@ import GHC.Generics (Generic)
 
 -- | Haskell type to convert to.
 data HaskellType = HaskellType
-    { htTitle :: Text
-    , htAtom  :: Double
-    , htCash  :: Bool
-    , htWords :: [Text]
-    , htBool  :: [Bool]
-    , htToday :: ZonedTime
-    , htFruit :: FruitInside
-    , htSize  :: SizeInside
+    { htTitle   :: Text
+    , htAtom    :: Double
+    , htCash    :: Bool
+    , htWords   :: [Text]
+    , htBool    :: [Bool]
+    , htToday   :: ZonedTime
+    , htInteger :: [Integer]
+    , htFruit   :: FruitInside
+    , htSize    :: SizeInside
     } deriving (Show, NFData, Generic)
 
 instance FromJSON HaskellType where
@@ -34,6 +35,7 @@ instance FromJSON HaskellType where
         <*> o .: "words"
         <*> o .: "bool"
         <*> o .: "today"
+        <*> o .: "ints"
         <*> o .: "fruit"
         <*> o .: "size"