+++ src/Toml/Parser/Core.hs
@@ -16,8 +16,7 @@ import Data.Void (Void)
 import Text.Megaparsec (Parsec, anySingle, errorBundlePretty, match, parse, satisfy, try, (<?>))
 import Text.Megaparsec.Char (alphaNumChar, char, digitChar, eol, hexDigitChar, space, space1,
                              string, tab)
-import Text.Megaparsec.Char.Lexer (binary, float, hexadecimal, octal, signed, skipLineComment,
-                                   symbol)
+import Text.Megaparsec.Char.Lexer (binary, float, hexadecimal, octal, signed, skipLineComment, symbol)
 import qualified Text.Megaparsec.Char.Lexer as L (lexeme, space)